Add isDelimiter function in Token.java and change codes in Main.java
This adds isDelimiter function which determines whether a read character
in string format includes a brace, a parenthesis, a comma, or a
semicolon.

This also changes the variable name of the read character in string
format in Main.java, since it may confuse with the read string buffer
variable.

There was a mistake in code where an incorrect directory was given as an
argument to File instance in Main.java, so it was corrected.