Add "ID" token, NullPointerException class and fileName variable
This adds identifier token name to the symbol table, as "ID".

This adds NullTokenException class which is thrown when a "NULL" token
is being put in the symbol table
If NullTokenException is thrown, the program shuts down.

This also separates input file name as a String variable "fileName" from
File instance in Main.java