fix: Make literal strings recognized correctly
Previously, literal string was not correctly recognized, in that the
lexer stopped reading characters after a double quote when it meets a
delimiter (ex. whitespace) and returned an exception.

This commit fixes the bug by changing the main function in Main.java, by
adding special conditions for literal strings and a boolean variable
called isLiteralString, which holds information of whether the lexer has
read a double quote previously.

Plus, there is a minor fix to token patterns, just by adding ( and ).