fix: Fix double quote reading bug
Previously, there was a bug where if the inputs received are after an
opening double quote, and when the lexer meets the last character that
is not a closing quote, the charAt() method returned a
NullPointerException.

This commit fixes it by implementing a mechanism to check if it is the
last character, when the program reads inputs after an opening double
quote.